Refactor the useQuote hook
